Refactored events out of atomicpuppy.py
Plus bonut code improvements
